-- Major code reorganization for ease of use. -- fixed grayplots, QC-rsFC, and deltaR plots -- grayplot and rsFC plot for denoised data are saved as .png in each subject's folder
Most functions which are not to be edited by the user are now in helpers.py.
helpers.py IS THE ONLY REQUIRED FILE, together with the main script which is data dependent.

The code can be launched from a simple jupyter notebook
[See notebook examples: CamCAN_movie.ipynb, CamCAN_movie_parallel.ipynb]
It would also be very easy to have a command line python script
[no example yet]
All user input is packaged into the config structure.

TODO:
-> check whether CIFTI format works, bug fix
-> make updated notebook for HCP IQ prediction